requirements_all.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
async-timeout==5.0.1
attrs==25.3.0
ccxt==4.4.95
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
dateparser==1.2.2
et_xmlfile==2.0.0
fonttools==4.59.0
frozenlist==1.7.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
matplotlib==3.9.4
multidict==6.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.3.0
propcache==0.3.2
pycares==4.9.0
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
ta==0.11.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
